Add setuptools to the installation requirements
`setuptools` is a dependency of this package, as it is used in
`version.py`. When building applications with this package as
dependency, the applications need to have `setuptools` as an additional
dependency, eventough it is this package's dependency.

By adding `setuptools` to the list of installation requirements, this
package becomes self-contained.